internal/featuretest: add RetryPolicy tests
Updates #1506

This is a followup to #1661. Move and consolidate the feature tests for
RetryPolicy acorss Ingress, IngressRoute and HTTPProxy.
